+++ Scripts/Bash/reissueFileVaultKey/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Reissue FileVault Key release history
 
+## 2.0.0 - 2023-02-23
+
+* Added better notifications and alerts.
+* Changed `displayDialog` icon parameter to allow icon resource names or ids
+* Changed `displayDialog` parameters position to make more sense.
+* Changed last supported OS version to macOS 12.
+
 ## 1.0.0 - 2023-02-22
 
 ### First version